Split notebooks
Refactor ddlite.py to submodules
Add pickler for extractions
